Incorporated increase quarter period algorithm to work with algorithm george
